N2-3 Barrel/Cap
Node 2/3 Barrel and Cap pieces.  Trying on the cap the idea of screw threads to make it easier (no threads on barrel, should be a snug fit). This method is far easier to deal with if it works, perhaps we put threads on the barrels too...  Inside of cap is down on print bed, open end is down on barrel (so the insides suck if need be).  Supports are probably a good idea. i'm using grid support with Cura, but it's tough to get out later (soluble support would rock here). I've left it with no ill effects so far. Still setup for 1/16" deep ring magnets, but i fixed the one radial port that was closed.  Suggest using that one for node 3 zenith, for N2 may also want this for zenith to support PMA3/IDA3/CVV.  Currently no grapple holes for this, will upgrade later with exterior details. (a cap could be made with a 1/8" hole in it for the PVGF at N3 zenith to fit into the 1/2" port hole. another upgrade for another time)