S0-S1/P1 models
Here's the inboard trusses previously not printed.  S1/P1 are identical, just print 2.  The end with the big block should be printed down. holes lines up with S0 for wire bundle passage and 2 side holes for magnets or rods as you see fit.  I'm about 33% through printing S0 (3.5 hrs expected...)  If you print with support, the trunnions should come out ok, otherwise expect to cut them off.  The hole in the top of the S1 is for a MicroServo9g, which should fit and allow wiring to fit into the wire tunnel.  I will need to modify the radiator beam/center radiator to accommodate the raised motor and still spin correctly.  I still need to work out the connection between S0/Lab (ala make my own MTSAS/LCA), as that's structural only.  I expect to modify the Lab to allow wiring to pass out the sides of the lab at the umbilicals/struts and up into the holes in the lower/forward side of the S0.