Metro board used by SARJs had bad pin 3.  Using pin 5 instead.
Debugging:
With D3 set as an encoder pin, it defaults to 0.  Other such declared pins default to 1.   In this cfg, jumper from 3.3v to Pin 3 does result in it toggling to 1.  Perhaps voltage from encoder isn't quite high enough to achieve this.  Note that any pin not declared as encoder pin (or otherwise manually set as input), will always have zero, even with jumper from 3.3v.

Component-level test (not installed in Mimic) shows good motor encoder reads for the pairs in code (2,5 for Port SARJ, 7,8 for Stbd SARJ).