2nd upload. Completes the base truss
Adds the 3 generic truss (no payload platform mounts), one with 2 mounts one with 4. Adds the solar array frame pieces to use with an acrylic tube for the mast (remember to cut off the blocks at the base or they will collide with the other side's motor). Adds the ITS-3 Alpha joint gear ring, and the other side piece (AGIS struts) to encapsulate the slip ring and connect to the 4 truss.  The Outboard truss support plugs into the end of the 6 truss, allowing for mounting on the end to the outboard slip ring (that mount is coming).  This completes the base truss components.