Updated the ITS-1 truss to this model.
The ITS-1 model didn't have the hole under one trunnion for the inboard support mounts (apologies). This fixes it, but, when you print it for the other side, you need to mirror it about a plane between the trunnions (perpendicular to the trunnions) so the hole ends up on the other side (both should be pointed down/toward the pressurized modules in the build). Found this as we're printing a 2nd model to test these prints.  Overall, I'm using cura, the horizontal expansion value set to keep build plane tolerances in check, printing mostly at 0.2mm layer heights, with supports.  Not all models are oriented for direct printing, use your best judgement, we'll eventually get notes for these up. (Or ask, we can help/suggest.)