2020-Endcap-tall component
This is the 20x20 extrusion endcap, on the outboard sides of the truss sliprings (the other side of the slipring connects to the Outboard-Truss-support.stl).  It's a tight fit for the slipring cylinder(needs to be), and the nub is to catch one of the holes in the slip ring flange to prevent it from rotating (I think I used the hole that makes the 2 wire bundles horizontal to each other, just for wire management/same length to go down the extrusion to the HDMI breakoutboard).  Hopefully you won't need to remove this after installation, but is possible.  The cap may be loose to the extrusion based on your printer's XY tolerances - various ways to fix this as you want (tape, glue, slicer settings, etc.).